Move processing exception to different package
